refactor category and product delete method

+++ backend/src/main/java/com/devsuperior/dscatalog/services/CategoryService.java
@@ -9,6 +9,7 @@ import com.devsuperior.dscatalog.services.exceptions.ResourceNotFoundException;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -68,15 +69,11 @@ public class CategoryService {
     }
 
     public void delete(Long id) {
-        if (repository.existsById(id)) {
-            try {
-                repository.deleteById(id);
-            }
-            catch (DataIntegrityViolationException e) {
-                throw new DatabaseException("Invalid request");
-            }
-        } else {
-            throw new ResourceNotFoundException("Id " + id + " not found");
+        try {
+            if (repository.existsById(id)) {repository.deleteById(id);}
+            else {throw new ResourceNotFoundException("Id not found " + id);}
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Integrity violation");
         }
     }
 }
