+++ backend/src/main/java/com/devsuperior/dscatalog/services/ProductService.java
@@ -9,11 +9,12 @@ import com.devsuperior.dscatalog.repositories.CategoryRepository;
 import com.devsuperior.dscatalog.repositories.ProductRepository;
 import com.devsuperior.dscatalog.services.exceptions.DatabaseException;
 import com.devsuperior.dscatalog.services.exceptions.ResourceNotFoundException;
+
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -67,15 +68,11 @@ public class ProductService {
     }
 
     public void delete(Long id) {
-        if (repository.existsById(id)) {
-            try {
-                repository.deleteById(id);
-            }
-            catch (DataIntegrityViolationException e) {
-                throw new DatabaseException("Invalid request");
-            }
-        } else {
-            throw new ResourceNotFoundException("Id " + id + " not found");
+        try {
+            if (repository.existsById(id)) {repository.deleteById(id);}
+            else {throw new ResourceNotFoundException("Id not found " + id);}
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Integrity violation");
         }
     }
